requirements.txt

# pip freeze > requirements.txt  - to get requirements form the project
# pip install -r requirements.txt  - to install requrements

attrs==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
h11==0.14.0
idna==3.10
iniconfig==2.0.0
outcome==1.3.0.post0
packaging==24.2
pluggy==1.5.0
PySocks==1.7.1
pytest==8.3.4
requests==2.32.3
selenium==4.29.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.12.2
urllib3==2.3.0
websocket-client==1.8.0
wsproto==1.2.0

mysql-connector-python~=9.2.0
python-dotenv~=1.0.1
pandas~=2.2.3
SQLAlchemy~=2.0.39
Faker~=37.0.2